Rename the whoStartsTheGame folder to whostart
