+++ src/main/java/jalau/at18/kingoftokyo/controller/WhoStartController.java
@@ -7,9 +7,7 @@ import jalau.at18.kingoftokyo.model.DiceFaceSameCounter;
 import jalau.at18.kingoftokyo.model.DiceHand;
 import jalau.at18.kingoftokyo.model.Monster;
 import jalau.at18.kingoftokyo.model.Player;
-
-import jalau.at18.kingoftokyo.view.whoStartsTheGame.WhoStartFrame;
-
+import jalau.at18.kingoftokyo.view.whostart.WhoStartFrame;
 
 public class WhoStartController {
     private WhoStartFrame whoStartFrame;
