+++ src/main/java/jalau/at18/kingoftokyo/view/whostart/PlayerNameLabel.java
@@ -1,4 +1,4 @@
-package jalau.at18.kingoftokyo.view.whoStartsTheGame;
+package jalau.at18.kingoftokyo.view.whostart;
 
 import javax.swing.*;
 import java.awt.*;
