+++ src/main/java/jalau/at18/kingoftokyo/view/whostart/WhoStartFrame.java
@@ -1,4 +1,4 @@
-package jalau.at18.kingoftokyo.view.whoStartsTheGame;
+package jalau.at18.kingoftokyo.view.whostart;
 
 import jalau.at18.kingoftokyo.controller.WhoStartRollDiceController;
 import jalau.at18.kingoftokyo.model.Player;
